+++ src/furniture/primitives/index.ts
@@ -8,9 +8,11 @@
 import type { ComponentType } from 'react'
 import type { ParamProps, PrimitiveKind } from '../types'
 import { AirconUnit } from './AirconUnit'
+import { AltarCabinet } from './AltarCabinet'
 import { Aquarium } from './Aquarium'
 import { AquariumStand } from './AquariumStand'
 import { Armchair } from './Armchair'
+import { Banquette } from './Banquette'
 import { BarCart } from './BarCart'
 import { BarStool } from './BarStool'
 import { BathroomSink } from './BathroomSink'
@@ -63,6 +65,7 @@ import { FloorLamp } from './FloorLamp'
 import { FloorMirror } from './FloorMirror'
 import { FloorSpeaker } from './FloorSpeaker'
 import { FloorVase } from './FloorVase'
+import { FlutedPartition } from './FlutedPartition'
 import { FoldingChair } from './FoldingChair'
 import { FruitBowl } from './FruitBowl'
 import { GarmentRack } from './GarmentRack'
@@ -145,6 +148,7 @@ import { WallShelf } from './WallShelf'
 import { WallTapestry } from './WallTapestry'
 import { Wardrobe } from './Wardrobe'
 import { WashingMachine } from './WashingMachine'
+import { WaterHeater } from './WaterHeater'
 import { WindowMeshScreen } from './WindowMeshScreen'
 import { WineCooler } from './WineCooler'
 
@@ -295,4 +299,8 @@ export const PRIMITIVE_COMPONENTS: Record<PrimitiveKind, PrimitiveComponent> = {
   WallDesk,
   LoftBed,
   TrestleDesk,
+  AltarCabinet,
+  Banquette,
+  WaterHeater,
+  FlutedPartition,
 }